refactor: Migrate Mimetype engine to directory engine.
